Update picker.py
issue #498 - Fix error in datepicker, when using two datepicker in one screen.  When the ValueErrror appears, the value for day is set to 1, since the function is run through twice. In the first run, the value for day is 31 and the ValueError is thrown. The selected value for day is only set in the second run.